+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
-## [0.0.1]
+## [0.1.0]
 
-- Initial Commit
+### Added
+
+- Toggle command to enable/disable the extension
+- `inactiveTimeout` setting to automatically close inactive tabs
+- `maxTabs` setting to limit the maximum number of open tabs